[#158936245] 'delete_vm' action return nil when instance is not found

+++ src/bosh-softlayer-cpi/softlayer/virtual_guest_service/softlayer_virtual_guest_service_delete.go
@@ -1,6 +1,16 @@
 package instance
 
+import bosherr "github.com/cloudfoundry/bosh-utils/errors"
+
 func (vg SoftlayerVirtualGuestService) Delete(id int, enableVps bool) error {
+	_, found, err := vg.softlayerClient.GetInstance(id, "id")
+	if err != nil {
+		return bosherr.WrapErrorf(err, "Failed to find SoftLayer VirtualGuest with id '%d'", id)
+	}
+	if !found {
+		return nil
+	}
+
 	if enableVps {
 		return vg.softlayerClient.DeleteInstanceFromVPS(id)
 	}
